pythonProject/requirements.txt
streamlit~=1.42.0
pandas~=2.2.3
python-dateutil~=2.9.0post0
openpyxl